refractored useModelIdRevisionIdFromModelOptions by removing useQueriedAddModelOptions used internally and adding the query llogic in itself
